ENH: Start adding PostgreSQL database + loads of other improvements
Add a separate API for a local PostgreSQL database for storing user data
and queries.
Add support for storing the user and query data into the database.
Add some comments + rename some parameters
Add a separate route for adding new users instead of doing it
automatically when a new user tries to log in and adjust the old login
route accordingly.