ENH: Remove direct db access from analysis tools and other fixes
Instead of using a target_id parameter in the analysis tasks, retrieve the
results to be analysed and send those to the tools as a separate parameter.

Separated split_document_set_by_facet as a tool of its own.

Add support for more generalized tools in the form of variable facet names.

Improve execute_async_tasks to run all the prerequisite queries for
analysis tasks in parallel.