ENH: Add support for POST search requests
search/ now accepts both GET and POST requests. POST request expect the
search parameters to be included as JSON in the message body. POST requests
can contain multiple searches in a single request, stored as a list of
JSON objects in the body. For single searches the result will be a single
JSON object, for multiple searches, a list of JSON objects.