ENH: Add a parameter for forcing a refresh to the results
Both search and analysis tasks now accept an additional body parameter:
'force_refresh'. If this is set to 'True', the task will be executed even
if there is already an existing result for the same task and the old result
is overwritten. The parameter itself is not stored to the internal db as
a part of the task parameters. Instead it is removed from the request as
soon as the task execution has started.