ENH: Refactor
Move the planning to dedicated class.
Move task generation and storing the results to a separate module.
Adjust the analysis_utils to accommodate the new planner.
Redo generate_time_series to get rid of the last instance of a utility
creating new tasks.
Fix shelltools to work with the new system